requirements.txt
pyro-ppl==1.7.0
torch
sklearn
pandas
numpy
tqdm
wordcloud
matplotlib
typer
selenium